webapp/backend/requirements.txt
# Core Django
Django>=4.2.7
python-dotenv>=1.0.0

# REST Framework
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
djangorestframework-knox>=4.2.0

# Chatbot Dependencies
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.24.0

# Development
gunicorn>=21.2.0
whitenoise>=6.5.0

# Utilities
python-dateutil>=2.8.2
Pillow>=10.0.0
